Fix phase-2 deploy script for ethers v5 (circomlibjs).
JsonRpcProvider lives under ethers.providers in v5; support both v5 and v6 APIs.

+++ contracts/src/lib/IncrementalMerkleTree.sol
@@ -9,7 +9,7 @@ pragma solidity ^0.8.24;
  */
 library IncrementalMerkleTree {
     uint256 internal constant DEPTH = 20;
-    uint256 internal constant MAX_LEAVES = 1 << DEPTH;
+    uint256 internal constant MAX_LEAVES = uint256(1) << DEPTH;
 
     struct Tree {
         uint256 nextIndex;
